feat RandomName: - member add/edit/view routers moved to router dir - url changes in navbar and home page
