+++ RandomName/templates/home.html
@@ -10,14 +10,14 @@
                     <div class="card-body">
                         <textarea id="card1" class="form-control text-area-dev" rows="1" readonly></textarea>
                         <br>
-                        <button class="btn btn-primary" onclick="fetchData('card1', 'random-dev-team-member')">Get DEV Member</button>
+                        <button class="btn btn-primary" onclick="fetchData('card1', 'members/random-dev-team-member')">Get DEV Member</button>
                     </div>
                 </div>
                 <div class="card text-center">
                     <div class="card-body">
                         <textarea id="card2" class="form-control text-area-dev" rows="1" readonly></textarea>
                         <br>
-                        <button class="btn btn-primary" onclick="fetchData('card2', 'random-qa-team-member')">Get QA Member</button>
+                        <button class="btn btn-primary" onclick="fetchData('card2', 'members/random-qa-team-member')">Get QA Member</button>
                     </div>
                 </div>
             </div>
@@ -27,7 +27,7 @@
                     <div class="card-body">
                         <textarea id="card3" class="form-control" rows="1" readonly></textarea>
                         <br>
-                        <button class="btn btn-primary" onclick="fetchData('card3', 'random-team-member')">Get Team Member</button>
+                        <button class="btn btn-primary" onclick="fetchData('card3', 'members/random-team-member')">Get Team Member</button>
                     </div>
                 </div>
                 
@@ -35,7 +35,7 @@
                     <div class="card-body">
                         <textarea id="card4" class="form-control" rows="1" readonly></textarea>
                         <br>
-                        <button class="btn btn-primary" onclick="fetchData('card4', 'random-tentative-team-member')">Get Co-Host</button>
+                        <button class="btn btn-primary" onclick="fetchData('card4', 'members/random-tentative-team-member')">Get Co-Host</button>
                     </div>
                 </div>
             </div>
