+++ RandomName/templates/navbar.html
@@ -12,12 +12,12 @@
             </ul>
             <ul class="navbar-nav">
                 <li class="nav-item active">
-                    <a class="nav-link" href="/add-member">Add Member</a>
+                    <a class="nav-link" href="/members/add-member">Add Member</a>
                 </li>
             </ul>
             <ul class="navbar-nav">
                 <li class="nav-item active">
-                    <a class="nav-link" href="/view-all">View All Members</a>
+                    <a class="nav-link" href="/members/view-all">View All Members</a>
                 </li>
             </ul>
         </div>
